Deprecated lt.dragas.birc.v3.core.Client
Due to the fact that it depends on deprecated lt.dragas.birc.v3.core.route.RouteGroup
class and has shoehorned multithreading in it, this class is unsustainable
and should let the implementations handle multithreading. It should be
replaced with lt.dragas.birc.v3.core.main.Client class, which doesn't have
internal multithreading support and does not not depend on deprecated RouteGroups,
while instead using a Router.